Update version to 4.0.15 in package.json; enhance precision handling for stats and actions in game logic, including support for crystals

+++ schemas/crystal.js
@@ -6,7 +6,7 @@ const CrystalSchema = z.object({
     slot: z.number().int().min(1).max(6),
     rarity: z.enum(['common', 'uncommon', 'rare', 'legendary', 'mythical', 'godlike']),
     stat: z.enum(['speed', 'health', 'attack', 'defense', 'criticalRate', 'criticalDamage', 'resist', 'focus']),
-    statValue: z.number().int(),
+    statValue: z.number()
 })
 
 module.exports = { CrystalSchema }
